refactor: move utility functions from lib.rs to utils.rs
Move non-public utility functions from lib.rs to a new utils.rs module to improve code organization and readability. This change:

- Creates a new utils.rs module for internal helper functions
- Moves all non-public functions from lib.rs to utils.rs
- Adds pub(crate) visibility to utility functions
- Updates imports and function calls in lib.rs
- Maintains all existing functionality while improving code structure

The refactor makes lib.rs more focused on the public API while keeping implementation details in a dedicated utils module.
